send user data to chat service when user is registered through oauth
