+++ user-service/src/services/authServices.ts
@@ -10,22 +10,7 @@ import httpStatusText from "../utils/httpStatusText";
 import compareHashedValues from "../utils/hashingUtils/compareHashedValues";
 import sanitizeUser from "../utils/sanitizeUser";
 import { TNewUser } from "../types";
-import publishMessage from "../utils/rabbitmqUtils/publishMessage";
-
-const sendUserDataToQueue = async (queueName: string, user: Partial<User>) => {
-  try {
-    const userData = {
-      userId: user.userId,
-      email: user.email,
-      username: user.username,
-    };
-
-    const stringifiedUserData = JSON.stringify(userData);
-    await publishMessage(queueName, stringifiedUserData);
-  } catch (error) {
-    throw error;
-  }
-};
+import sendUserDataToQueue from "../utils/rabbitmqUtils/sendUserDataToQueue";
 
 const signupService = async (userData: TNewUser) => {
   try {
